refactor: reorganize imports and move GlassExpert class to its own file
